Allow to uncrustify files with specified file extensions only.
We do not want to run uncrustify on documentation or other text files
that do not contain code. It is now possible to specify whether only
files of specified type are parsed and which extensions are checked.

By default it now only checks .{c, cpp, h, hpp} files for style
violations.