Support source code filenames with spaces
If the names of source code files checked by pre-commit-uncrustify
contain spaces or other special characters, the patch file must be
adapted accordingly.

If the diff patch file contains headers like:

--- a/path with spaces/file with spaces.cpp    [date]
+++ b/path with spaces/file with spaces.cpp    [date]

the "git apply" command will not be able to apply the patch, as it will
be confused by the spaces.

Fortunately it is possible to produce applicable patches in such cases,
by quoting the file names accordingly:

--- "a/path with spaces/file with spaces.cpp"  [date]
+++ "b/path with spaces/file with spaces.cpp"  [date]

It is important to not only quote the file names, but the leading "a/"
and "b/" prefixes must be qouted along with the filenames, like in the
example above.