Use command to check for uncrustify binary instead of checking the path.
It makes more sense to check whether uncrustify is installed with
'command -v' than checking whether the path is executable. This way we
can simply specify 'uncrustify' as the executable in the configuration
and do not have to enter a full path that varies from user to user. This
means less required configuration by the user when the hooks are
installed which is always a good thing.