Do not exit when files are skipped during installation.
Unfortunately 'cp -i' returns with non-zero when a file is being skipped
by the user. As we run the install script with 'set -e' and exit on
errors, the script aborts with the first skipped file. Explicitly ignore
the cp return value instead. This means we do not exit on actual errors
for that call anymore either which is not ideal but okay for now.
Ideally we would do a separate check here.