Add initial README.md with example pictures.
This is a first step towards a readme that makes it more clear what this
repository is about. It comes with some example pictures and workflow to
show how the Uncrustify hook works. Otherwise, when opening the
repository for the first time it is not very clear what to expect from
the hooks.