Use Uncrustify's automatic language detection.
If no language option is specified, uncrustify detects the language
itself (possibly from the file extension). This seems to work fine from
my experience. It is now the default option which means one less item
the user has to configure before using the hook. And having to do less
manual configuration is always a good thing.

It can still be manually set if the detection works incorrectly or the
user does not trust the automatic detection.