Generate a more readable name for patch files.
Rather than using the somewhat random looking seconds of the current
time for the generated uncrustify patches, use a more readable time
string in the form of 'yyyy-mm-dd_hhmmss'. We are not trying to mask
anything here and the goal was simply to avoid collisions. This is still
the case but now it is a lot easier to go back and look up a generated
patch by name rather than having to look at the original script
output or last edited time on the files.