Reduce filesize of example pictures.
The example pictures are now compressed with the lossy compression
algorithm from tinypng.com. It does not show as a visible difference,
but it saves 74% of the overall sizes and reduces the added sizes from
245.3KB to 63.5KB. This should reduce the time it takes to load the
pictures when viewing the readme.

The filesizes may not seem like a lot originally, but on the slow local
internet connection here they still always took a few visible moments
to load.