Remove git default hook from repo.
I am not quite sure what the licensing situation is with the example
pre-commit hook that comes shipped with git. Git itself is licensed with
a GNU license, which at the very least requires to be mentioned when the
sources are used. In this case it is easier to just remove the example
hook and any mention of it. It comes pre-installed with git anyway, so
if we want to, we can simply update the installer to set it up
when the hooks from this repository are installed.