Remove default git hook from pre-commit.
We missed to remove it from the base hook before. It is not part of the
repository anymore and should not run by default.
If we do want to run it, then we have to undo this change again and
update the installer so that it renames the example hook that comes with
every git installation.