Use config file for install_hook.sh configuration.
Currently we have all configuration inside the hooks and scripts
themselves. This is somewhat annoying though as any changes are visible
in the repository. A better approach is to use separate config files
that are not tracked by git and that can be used as necessary.

We start with install_hooks as it is the most simple script. We add a
install_hooks.example.cfg example documentation file, and expect a
install_hooks.cfg with the actual configuration.

Note that with this commit we do not actually install the config file(s)
themselves yet.