Determine hooks to be installed from pre-commit.cfg.
This avoids a bit of duplication. The pre-commit hook needs to specify
which hooks are going to be run. It makes sense to use the hooks set in
there to determine which ones to install. This avoids us having to set
up two basically identical config files for both install_hook.sh and
pre-commit.