Only install example hook config if actual config does not exist.
There is no point in copying the example config if a regular config is
already being copied over as well. Also, if we do install the example
config we can just as well install it with the 'example' in the name so
that it is actually used and can simply be updated after it is
installed.